테스트 코드 개선
